[#173] Refactor: 챌린지 추천 API 수정
