+++ src/main/java/umc/GrowIT/Server/service/diaryService/DiaryCommandService.java
@@ -1,5 +1,6 @@
 package umc.GrowIT.Server.service.diaryService;
 
+import umc.GrowIT.Server.web.dto.ChallengeDTO.ChallengeResponseDTO;
 import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryRequestDTO;
 import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryResponseDTO;
 
@@ -14,4 +15,6 @@ public interface DiaryCommandService {
     public DiaryResponseDTO.VoiceChatResultDTO chatByVoice(DiaryRequestDTO.VoiceChatDTO request, Long userId);
 
     public DiaryResponseDTO.SummaryResultDTO createDiaryByVoice(DiaryRequestDTO.SummaryDTO request, Long userId);
+
+    DiaryResponseDTO.AnalyzedDiaryResponseDTO analyze(Long diaryId);
 }
