+++ src/main/java/umc/GrowIT/Server/service/diaryService/DiaryCommandServiceImpl.java
@@ -1,16 +1,18 @@
 package umc.GrowIT.Server.service.diaryService;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import umc.GrowIT.Server.apiPayload.code.status.ErrorStatus;
-import umc.GrowIT.Server.apiPayload.exception.DiaryHandler;
-import umc.GrowIT.Server.apiPayload.exception.UserHandler;
+import umc.GrowIT.Server.apiPayload.exception.*;
 import umc.GrowIT.Server.converter.DiaryConverter;
-import umc.GrowIT.Server.domain.Diary;
-import umc.GrowIT.Server.domain.User;
+import umc.GrowIT.Server.domain.*;
+import umc.GrowIT.Server.repository.ChallengeKeywordRepository;
+import umc.GrowIT.Server.repository.ChallengeRepository;
+import umc.GrowIT.Server.repository.KeywordRepository;
 import umc.GrowIT.Server.repository.UserRepository;
 import umc.GrowIT.Server.repository.diaryRepository.DiaryRepository;
 import umc.GrowIT.Server.web.dto.DiaryDTO.DiaryRequestDTO;
@@ -21,18 +23,26 @@ import umc.GrowIT.Server.web.dto.OpenAIDTO.Message;
 
 import java.time.LocalDate;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class DiaryCommandServiceImpl implements DiaryCommandService{
     private final DiaryRepository diaryRepository;
     private final UserRepository userRepository;
+    private final ChallengeKeywordRepository challengeKeywordRepository;
+    private final KeywordRepository keywordRepository;
+    private final ChallengeRepository challengeRepository;
 
     @Value("${openai.model1}")
     private String chatModel;
 
     @Value("${openai.model2}")
     private String summaryModel;
+
+    @Value("${openai.model2}")
+    private String openAIModel;
 
     @Value("${openai.api.url}")
     private String apiURL;
@@ -201,5 +211,186 @@ public class DiaryCommandServiceImpl implements DiaryCommandService{
         return DiaryConverter.toSummaryResultDTO(diary);
     }
 
+    // 일기 분석
+    @Override
+    public DiaryResponseDTO.AnalyzedDiaryResponseDTO analyze(Long diaryId) {
+        // 1. 일기 조회
+        Diary diary = diaryRepository.findById(diaryId)
+                .orElseThrow(() -> new DiaryHandler(ErrorStatus.DIARY_NOT_FOUND));
 
+
+        // 2. 예외 체크
+        // 일기 분석은 1번만 가능
+        if(diary.getDiaryKeywords() != null && !diary.getDiaryKeywords().isEmpty()) {
+            throw new DiaryHandler(ErrorStatus.ANALYZED_DIARY);
+        }
+        // 당일의 일기만 분석 가능
+        if (!diary.getDate().isEqual(LocalDate.now())) {
+            throw new DiaryHandler(ErrorStatus.DIARY_NOT_TODAY);
+        }
+
+
+        // 3. DB에서 감정들 조회
+//        List<String> emotions = keywordRepository.findAll()
+//                .stream()
+//                .map(emotion -> emotion.getName())
+//                .toList();
+//        log.info(emotions.toString());
+
+        // DB 설정이 제대로 안되었기 때문에 테스트용으로 DB ID 1~5까지의 감정 사용
+        // TODO 이후 위의 코드로 수정 필요
+        List<String> emotions = new ArrayList<>();
+        emotions.add("행복한");
+        emotions.add("슬픈");
+        emotions.add("외로운");
+        emotions.add("불안한");
+        emotions.add("기쁜");
+
+
+        // 4. OpenAI API 호출하여 감정 반환 & 반환받은 감정 체크 (3개인지, 중복안되었는지, DB와동일한지)
+        List<Keyword> analyzedEmotions = openAIAnalyzeDiary(diary.getContent(), emotions.toString());
+        log.info("[최종 분석된 감정] : " + analyzedEmotions.stream().map(Keyword::getName).toList().toString());
+
+
+        // 5. 일기와 감정키워드 연관관계 설정
+        setDiaryKeyword(diary, analyzedEmotions);
+
+
+        // 6. 감정 키워드 2개 랜덤 선택
+        List<Keyword> selectedKeyword = selectRandomKeyword(analyzedEmotions);
+        log.info("[선택된 감정] : " + selectedKeyword.stream().map(Keyword::getName).toList().toString());
+
+
+        // 7. 연관관계 맺고 있는 챌린지 중에서 각 랜덤으로 1개씩 선택 (=데일리 챌린지)
+        List<Challenge> dailyChallenges = selectDailyChallenge(selectedKeyword);
+        log.info("[데일리 챌린지] : " + dailyChallenges.stream().map(Challenge::getTitle).toList().toString());
+
+
+        // 8. 7)에서 선택되지 않은 챌린지들 중에서 랜덤으로 1개 선택 (=랜덤 챌린지)
+        Challenge randomChallenge = challengeRepository.findRandomRemainingChallenge(dailyChallenges);
+        log.info("[랜덤 챌린지] : " + randomChallenge.getTitle());
+
+
+        // 9. converter작업
+        return DiaryConverter.toAnalyzedDiaryDTO(analyzedEmotions, dailyChallenges, randomChallenge);
+    }
+
+
+    // 일기 분석 (일기 -> 감정)
+    private List<Keyword> openAIAnalyzeDiary(String diaryContent, String emotions) {
+
+        // 프롬프트
+        String prompt = String.format("""
+            다음 일기를 분석하여 %s 감정들 중 가장 해당되는 감정을 3개 선택해주세요.
+            
+            일기 내용: %s
+            
+            반환 결과는 다른 말 없이 감정들만을 []로 둘러싸서 반환해 주세요.
+            ex) [감정1, 감정2, 감정3]
+            """, emotions, diaryContent);
+
+
+        // API 요청 생성
+        ChatGPTRequest gptRequest = new ChatGPTRequest(openAIModel, prompt);
+
+        // API 요청 및 응답 처리
+        ChatGPTResponse chatGPTResponse = template.postForObject(apiURL, gptRequest, ChatGPTResponse.class);
+        if (chatGPTResponse == null || chatGPTResponse.getChoices().isEmpty()) {
+            throw new OpenAIHandler(ErrorStatus.GPT_RESPONSE_EMPTY);
+        }
+        String result = chatGPTResponse.getChoices().get(0).getMessage().getContent();
+        log.info("[GPT 결과] : " + result);
+
+
+        // 응답 결과 정리 ([]를 제거하고 각 감정을 분리)
+        result = result.replace("[", "").replace("]", "").trim(); // 괄호 제거 후 양 옆 공백 제거
+        log.info("[결과정리1] : " + result);
+
+        List<String> emotionsList = Arrays.asList(result.split("\\s*,\\s*")); // , 앞뒤 공백 제거하고 감정들을 나누기
+        log.info("[결과정리2] : " + emotionsList.toString());
+
+
+        // 결과 체크
+        // 3개인지 체크
+        if (emotionsList.size() != 3) {
+            throw new OpenAIHandler(ErrorStatus.EMOTIONS_COUNT_INVALID);
+        }
+
+        // 중복 체크
+        Set<String> uniqueEmotions = new HashSet<>(emotionsList);
+        if (uniqueEmotions.size() != emotionsList.size()) {
+            throw new OpenAIHandler(ErrorStatus.EMOTIONS_DUPLICATE);
+        }
+
+        // DB에 존재하는 감정인지 체크
+        List<Keyword> emotionKeywords = uniqueEmotions.stream()
+                .map(emotion -> keywordRepository.findByName(emotion)
+                        .orElseThrow(() -> new KeywordHandler(ErrorStatus.KEYWORD_NOT_FOUND)))
+                .toList()
+                ;
+
+        return emotionKeywords;
+    }
+
+    // 일기와 감정키워드 연관관계 설정
+    private void setDiaryKeyword(Diary diary, List<Keyword> analyzedEmotions) {
+        // Diary의 diaryKeywords 리스트가 null이라면 초기화
+        if (diary.getDiaryKeywords() == null) {
+            diary.setDiaryKeywords(new ArrayList<>());
+        }
+
+        // 각 Keyword에 대해 DiaryKeyword 추가
+        for (Keyword keyword : analyzedEmotions) {
+            DiaryKeyword diaryKeyword = DiaryKeyword.builder()
+                    .diary(diary)
+                    .keyword(keyword)
+                    .build();
+
+            diary.getDiaryKeywords().add(diaryKeyword); // Diary에 DiaryKeyword 추가
+        }
+
+        diaryRepository.save(diary);
+    }
+
+
+    // 감정키워드 랜덤 2개 선택
+    private List<Keyword> selectRandomKeyword(List<Keyword> analyzedEmotions) {
+        Random random = new Random();
+
+        List<Keyword> editableList = new ArrayList<>(analyzedEmotions);
+        Collections.shuffle(editableList, random);
+
+        return editableList.subList(0, 2);
+    }
+
+
+    // 데일리챌린지 랜덤 2개 선택
+    private List<Challenge> selectDailyChallenge(List<Keyword> selectedKeyword) {
+        Random random = new Random();
+        List<Challenge> selectedChallenges = new ArrayList<>();
+
+        // 전달받은 selectedKeyword를 통해 연관된 챌린지들 조회
+        for (Keyword keyword : selectedKeyword) {
+            // 매핑 테이블을 통해 해당 키워드와 연관된 챌린지들을 가져오기
+            List<ChallengeKeyword> challengeKeywords = challengeKeywordRepository.findByKeywordWithChallenge(keyword);
+
+            // 연관된 챌린지들
+            List<Challenge> relatedChallenges = challengeKeywords.stream()
+                    .map(ChallengeKeyword::getChallenge)
+                    .filter(challenge -> !selectedChallenges.contains(challenge))  // 이미 선택된 챌린지 제외
+                    .collect(Collectors.toList());
+
+            if (relatedChallenges.isEmpty()) {
+                throw new ChallengeHandler(ErrorStatus.RELATED_CHALLENGE_NOT_FOUND);
+            }
+
+            Collections.shuffle(relatedChallenges, random);
+            Challenge selectedChallenge = relatedChallenges.get(0);
+            selectedChallenges.add(selectedChallenge);
+
+            log.info("[" + keyword.getName() + "]과 연관된 [" + selectedChallenge.getTitle() + "]을 데일리 챌린지로 선택");
+        }
+
+        return selectedChallenges;
+    }
 }
