+++ src/main/java/umc/GrowIT/Server/web/dto/ChallengeDTO/ChallengeRequestDTO.java
@@ -31,14 +31,4 @@ public class ChallengeRequestDTO {
         private List<Long> challengeIds;
         private UserChallengeType dtype;
     }
-
-
-    @Data
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class RecommendChallengesRequestDTO {
-        @NotNull
-        String diaryContent; //일기 내용
-    }
 }
